ex_banking v0.1.0 ExBanking.Account

The Account module is responsible for handling interactions between the ExBanking functions and the GenServer implemented in AccountServer.

It handles some business logic while also translating the low-level messages from AccountServer to business messages that ExBanking can actually return.

Functions

create_user(user)

create_user(user :: String.t()) :: {:ok, pid()} | {:error, :user_already_exists}

Creates a new AccountServer process using the given name (case-sensitive string), unless there is already a process using the same user name.

Returns :ok

Examples

iex> Account.create_user("Anna")
:ok

iex> Account.create_user("Olivia")
...> Account.create_user("Olivia")
{:error, :user_already_exists}

deposit(user, amount, currency, account_server_module \\ AccountServer)

deposit(
  user :: String.t(),
  amount :: integer(),
  currency :: String.t(),
  account_server_module :: atom()
) :: {:ok, integer()} | {:error, atom()}

Adds the given amount of money from the given currency to the specified user account.

Returns {:ok, balance}

Examples

iex> Account.create_user("Susan")
...> Account.deposit("Susan", 1042, "USD")
{:ok, 1042}

iex> Account.deposit("Santa Claus", 1050, "USD")
{:error, :user_does_not_exist}

get_balance(user, currency, account_server_module \\ AccountServer)

get_balance(
  user :: String.t(),
  currency :: String.t(),
  account_server_module :: atom()
) :: {:ok, integer()} | {:error, atom()}

Gets the current amount of money a user has on the given currency.

Returns {:ok, balance}

Examples

iex> Account.create_user("Tod")
iex> Account.get_balance("Tod", "USD")
{:ok, 0}

iex> Account.get_balance("Santa Claus", "USD")
{:error, :user_does_not_exist}

send(from_user, to_user, amount, currency, account_server_module \\ AccountServer)

send(
  from_user :: String.t(),
  to_user :: String.t(),
  amount :: integer(),
  currency :: String.t(),
  account_server_module :: atom()
) :: {:ok, integer(), integer()} | {:error, atom()}

Transfers an amount of money of a given currency from one user account to another. It basically runs a withdraw followed by a deposit, handling any rollback in case a operation fails.

Returns {:ok, sender_balance, receiver_balance}

Examples

iex> Account.create_user("Denis")
...> Account.create_user("Camila")
...> Account.deposit("Denis", 150, "USD")
...> Account.send("Denis", "Camila", 50, "USD")
{:ok, 100, 50}

withdraw(user, amount, currency, account_server_module \\ AccountServer)

withdraw(
  user :: String.t(),
  amount :: integer(),
  currency :: String.t(),
  account_server_module :: atom()
) :: {:ok, integer()} | {:error, atom()}

Removes the given amount of money from the given currency from the specified user account.

Returns {:ok, balance}

Examples

iex> Account.create_user("Marco")
...> Account.deposit("Marco", 100, "USD")
...> Account.withdraw("Marco", 50, "USD")
{:ok, 50}

iex> Account.withdraw("Santa Claus", 1050, "USD")
{:error, :user_does_not_exist}