ex_banking v0.1.0 ExBanking.AccountServer

The AccountServer module is a GenServer that is responsible for one user account at a time (multiple of them are spawned under the AccountsSupervisor).

It handles most of the concurrency-dependent code and is in a considerably lower-level than the Account module for example.

Functions

call(user, action_options, options \\ [])

call(user :: String.t(), action_options :: tuple(), options :: list()) ::
  {:ok, number()} | {:error, atom(), atom()}

Calls the AccountServer process of the given user and returns whatever the the process returns from the given message.

Examples

iex> AccountServer.spawn_account("Lara")
...> AccountServer.call("Lara", {:get_balance, "USD"})
{:ok, 0}

iex> AccountServer.spawn_account("Francis")
...> AccountServer.call("Francis", {:deposit, 100, "USD"})
{:ok, 100}

iex> AccountServer.spawn_account("Jeff")
...> AccountServer.call("Jeff", {:deposit, 100, "USD"})
...> AccountServer.call("Jeff", {:withdraw, 10, "USD"})
{:ok, 90}

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

handle_call(msg, from, state)

Callback from GenServer implementing the withdraw logic. Examples on the AccountServer.call function documentation.

spawn_account(user)

spawn_account(user :: String.t()) ::
  {:ok, pid()} | {:error, {:already_started, pid()}}

Spawns a new AccountServer GenServer under the AccountsSupervisor.

Returns {:ok, #PID<...>}

Examples

iex> {:ok, pid} = AccountServer.spawn_account("Sophie")
...> is_pid(pid)
true

iex> AccountServer.spawn_account("Mary")
...> {:error, {:already_started, pid}} = ExBanking.AccountServer.spawn_account("Mary")
...> is_pid(pid)
true

start_link(process_name)