ExBanking.Money.Amount (ex_banking v0.1.0)

To avoid floating point madness, currency is expressed in cents

Functions

from(amount)

Specs

from(amount :: float() | integer()) :: integer()

Normalizes amount coming in different formats Amounts is considered to be given always as currency, so 10 euro and 10.0 euro are the same value and will be represented internally as 1000

iex> alias ExBanking.Money.Amount
iex> Amount.from(10.0)
1000
iex> Amount.from(10)
1000

to_presentation(amount)

Specs

to_presentation(amount :: integer()) :: float()

Present the given currency value as a float

iex> alias ExBanking.Money.Amount
iex> Amount.to_presentation(1000)
10.0