ExBanking.User.UserStore (ex_banking v0.1.0)

Storage for User information (queue)

Types

user_data()

Specs

user_data() :: {binary(), integer()}

Functions

create(username)

Specs

create(username :: String.t()) :: :ok

Creates a new user or replaces the existing one. This is wanted, to allow UserWorker to initialize its queue in case of abnormal restarts

decrease(username, count \\ 1)

Specs

decrease(username :: String.t(), count :: number()) :: integer() | boolean()

flush()

Specs

flush() :: :ok

increase(username, count \\ 1)

Specs

increase(username :: String.t(), count :: number()) :: integer() | boolean()

init()

Specs

init() :: :ok

lookup(username)

Specs

lookup(username :: String.t()) :: {:ok, user_data()} | {:error, :not_found}

lookup_count(username)

Specs

lookup_count(username :: String.t()) :: {:ok, integer()} | {:error, :not_found}

Gets the queue size for a user

set_count(username, num)

Exported mostly for test

threshold()