ex_banking v0.1.0 ExBanking

The ExBanking module serves as an entrypoint for the application, it's APIs are the only things meant to be directly exposed to the outside world.

It's a very thin layer of the application, mostly just calling Account module methods instead of doing much on it's own.

Types

banking_error()

banking_error() ::
  {:error,
   :wrong_arguments
   | :user_already_exists
   | :user_does_not_exist
   | :not_enough_money
   | :sender_does_not_exist
   | :receiver_does_not_exist
   | :too_many_requests_to_user
   | :too_many_requests_to_sender
   | :too_many_requests_to_receiver}

Functions

create_user(user)

create_user(user :: String.t()) :: :ok | banking_error()

Creates a new user with a given name (case-sensitive string), unless there is already an user with the same name.

Returns :ok

Examples

iex> ExBanking.create_user("John")
:ok

iex> ExBanking.create_user("Robert")
...> ExBanking.create_user("Robert")
{:error, :user_already_exists}

deposit(user, amount, currency)

deposit(user :: String.t(), amount :: number(), currency :: String.t()) ::
  {:ok, new_balance :: number()} | banking_error()

Adds the given amount of money from the given currency to the specified user account.

Returns {:ok, balance}

Examples

iex> ExBanking.create_user("Evan")
...> ExBanking.deposit("Evan", 10.42, "USD")
{:ok, 10.42}

iex> ExBanking.deposit("Santa Claus", 10.50, "USD")
{:error, :user_does_not_exist}

get_balance(user, currency)

get_balance(user :: String.t(), currency :: String.t()) ::
  {:ok, balance :: number()} | banking_error()

Gets the current amount of money a user has on the given currency.

Returns {:ok, balance}

Examples

iex> ExBanking.create_user("Elton")
...> ExBanking.get_balance("Elton", "USD")
{:ok, 0.00}

iex> ExBanking.get_balance("Santa Claus", "USD")
{:error, :user_does_not_exist}

send(from_user, to_user, amount, currency)

send(
  from_user :: String.t(),
  to_user :: String.t(),
  amount :: number(),
  currency :: String.t()
) ::
  {:ok, from_user_balance :: number(), to_user_balance :: number()}
  | banking_error()

Transfers an amount of money of a given currency from one user account to another.

Returns {:ok, sender_balance, receiver_balance}

Examples

iex> ExBanking.create_user("Jose")
...> ExBanking.create_user("Amelia")
...> ExBanking.deposit("Jose", 150.0, "USD")
...> ExBanking.send("Jose", "Amelia", 50.0, "USD")
{:ok, 100.0, 50.0}

withdraw(user, amount, currency)

withdraw(user :: String.t(), amount :: number(), currency :: String.t()) ::
  {:ok, new_balance :: number()} | banking_error()

Removes the given amount of money from the given currency from the specified user account.

Returns {:ok, balance}

Examples

iex> ExBanking.create_user("Terry")
...> ExBanking.deposit("Terry", 150.50, "USD")
...> ExBanking.withdraw("Terry", 25.50, "USD")
{:ok, 125.0}

iex> ExBanking.withdraw("Santa Claus", 10.50, "USD")
{:error, :user_does_not_exist}